# 🚀 Fix #1: Adaptive Game Loop Frequency Optimization